requirements.txt
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
jinja2>=3.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
markdown>=3.5.0
httpx>=0.27.0
pygments>=2.17.0
